refactor: moving gcp/vertex specific code to vertex runner and tests
