+++ tests/unit_tests_old/tests_launch/test_launch_docker.py
@@ -13,7 +13,6 @@ from wandb.sdk.launch._project_spec import (
     fetch_and_validate_project,
 )
 from wandb.sdk.launch.builder.build import (
-    construct_gcp_image_uri,
     docker_image_exists,
     generate_dockerfile,
     get_base_setup,
@@ -212,30 +211,6 @@ def test_buildx_not_installed(
     )
 
     assert "RUN WANDB_DISABLE_CACHE=true" in dockerfile
-
-
-def test_gcp_uri(test_settings, live_mock_server, mocked_fetchable_git_repo):
-    api = wandb.sdk.internal.internal_api.Api(
-        default_settings=test_settings, load_settings=False
-    )
-    test_spec = {
-        "uri": "https://wandb.ai/mock_server_entity/test/runs/1",
-        "entity": "mock_server_entity",
-        "project": "test",
-        "cuda": None,
-        "resource": "local",
-        "resource_args": {},
-    }
-    test_project = create_project_from_spec(test_spec, api)
-    test_project = fetch_and_validate_project(test_project, api)
-
-    uri = construct_gcp_image_uri(
-        test_project, "test-repo", "test-project", "test-registry"
-    )
-    assert (
-        "test-registry/test-project/test-repo/wandb.ai__mock_server_entity__test__runs__1"
-        in uri
-    )
 
 
 def test_docker_image_exists(
